refactor: split app coordinator actions
